+++ src/pages/feed/css/FeedList/FeedPopularyRanking.module.css
@@ -1,0 +1,44 @@
+.rankingContainer {
+    margin-top: 20px;
+}
+
+.rankingTitle {
+    text-align: center;
+    margin-bottom: 20px;
+    font-size: 24px;
+    font-weight: bold;
+}
+
+.feedListItem {
+    padding: 15px;
+    margin-bottom: 10px;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    transition: box-shadow 0.3s ease;
+}
+
+.feedListItem:hover {
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
+}
+
+.feedInfo {
+    display: flex;
+    align-items: center;
+}
+
+.feedInfo span {
+    margin-right: 10px;
+    cursor: pointer;
+}
+
+.feedTitle {
+    font-size: 18px;
+    font-weight: bold;
+}
+
+.feedPopularityBadge {
+    background-color: #6c757d;
+    color: white;
+    font-size: 14px;
+    padding: 5px 10px;
+}
